refactor : 폴더 구조 변경으로 인한 위치 변경
body : 인증 인스턴스를 위한 header 설정
